[svn] Changed the debug callback handler function to accept a syslog-style priority as well:   typedef void (*GGZDebugHandlerFunc)(int priority, const char *msg);
This is the only change to the interface, although some changes to msg.c
were needed as well, of course.  The priority associated with the message
depends on how the message was generated:
  ggz_debug and _ggz_debug => LOG_DEBUG
  _ggz_msg => LOG_NOTICE (?)
  ggz_error_sys and ggz_error_msg => LOG_ERR
  ggz_error_sys_exit and ggz_error_msg_exit => LOG_CRIT

Libggz is not capable of providing nearly all of the features currently
used by ggzd for logging/debugging.  The only thing extra feature ggzd
uses is a separation of "logging" and "debugging" messages.