[svn] A quick conversion to get the GGZCards server to use ggz_debug and friends instead of ggzdmod_log.  A debug handler function is registered with libggz, and calls ggzdmod_log to pass the output on to GGZ.
This system makes for very convenient logging, but at the front (easy to
code) and the back (easy to read and control).  But it does have a price:
for instance a message logged by the AI goes through (at least) 4 different
printf-style formatting runs: 1 when ggz_debug is called in the AI process,
1 when ggzcards reads this from the error socket, 1 when ggzcards passes it
to ggzdmod, and 1 when ggzd puts the final thing together to send to the
debug file.  Ouch!