[svn] - when spectators leave/join, send their specific events, not player leave/join   events
This is starting to cause headache.
Spectators are separate entities, so they're not on the table. This means that
people should be allowed to watch several games while playing or not
playing one on their own. The GGZ core client thus doesn't care about how many
games a player watches in one room.
Hm...