[svn] - Debconf support for the GGZ server
This is how it works:
Whenever someone installs ggzd (or reconfigures it via dpkg-reconfigure ggzd),
a dialog box pops up and asks for the user's admin name and email address
(which are then injected into ggzd's config file), and for the games to be
installed.

Two notes:
- the game menu is vapourware currently, maybe it doesn't even belong here, but
  it's a start.
- the priority of the questions is 'medium'. Since too many questions during
  system installations suck, this could be set to 'low'. But only if folks
  complain (after all, there's 'high' and 'critical' at the other end).

Debconf can be fully localized, the template files must then be called
ggzd.templates.$LANG.

Note that in the future, a distributor-independent configuration panel, maybe
based on ncurses or something, would be cool.