[svn] - Powerful translation backend for grubby using google's translation service. - Initial code written by chimera
It is intended to work like
	grubby translate "oh my god this is a whole sentence" en-es

All translations are cached (there's a bug but this is fixed soon).
The only thing left to do is to recognize quotes from the input. On the command
line, this works already.

For single words, the traditional translator will still be faster I guess.