[svn] Instead of <JOINSPECTATOR/> and <LEAVESPECTATOR/>, make the XML protocol use <JOIN SPECTATOR='true'/> and <LEAVE SPECTATOR='true'/>.
This gives sufficient backwards-compatability: now if you try to "watch" a
game on a server that doesn't support spectators, you'll just join the
game as a player.  The other way, if you did this the client would become
fubar.

Note that this *does* break compatability with the CVS code of the past
several days.  But this is a small price to pay.