[svn] Doh!  We have to make sure to skip over the empty spectator seats when iterating over spectators.
In game_do_move, there's a potential problem because we return -1 if
there's ever a network error.  But at the end of the function is where the
game logic (incrementing the move, etc.) takes place.  So although this
shouldn't be a fatal error (it's just one player gone missing), the table
cannot recover.