[svn] Problem: when spectators join and leave tables, the other clients do not know what table those players are at.
Solution: ggzcore must track the list of spectators for each table, so
that when a table update is sent it can properly remove/add players from
the table.  This is, needless to say, a rather involved process.

There's a significant hack involved: the maximum number of spectators is
hard-coded at 8.  The "correct" thing to do would be to use a GGZList
which *only* has elements for the existing spectators.  That way the
list could grow arbitrarily.  All in good time...