[svn] in remove_table(), don't destroy the table data until after we've removed all of the players and spectators from it.
The specific problem we run into if we do it the other way is that after
the gametype is reset, ggzd doesn't know whether spectators are allowed
at the table.  This means, first of all, that spectator_seats_num() could
fail in remove_table if it's called before the gametype is reset (easy to
work around).  But it also means that when we send the <SPECTATOR>
updates to the clients, we won't know the number of spectators there
either (causing the updates to be interpreted incorrectly at the client
end).