[svn] Removed the spectator seat limit from GGZdMod.
Now the spectator seat list is just allocated dynamically.  Empty
spectator seats don't have an entry in the list, so there is a much
smaller memory overhead.

The function ggzdmod_set_max_num_spectators is no longer needed - the max
number of spectators (i.e. one higher than the highest spectator
number) is tracked internally by ggzdmod, and is still available through
ggzdmod_get_max_num_spectators.  This means iterating through spectators
is faster, so there is a smalle overhead there as well.

Also, if a spectator or player join fails in ggzdmod at the game end,
instead of failing silently an error code will be sent back to
ggzdmod-ggz.  The mechanism for this was already in place, but before if
there was an error ggzdmod-game would just do nothing.

This new system should be much more flexible, and should last us for some
time.