[svn] Converted all malloc (and friends) routines to ggz_malloc (and friends) ones.
This is a change I've been thinking about/working on for a while.  In the
process, I've fixed several problems and converted several
heap-allocation routines to stack-allocation ones.  All in all, it should
be a big step forward.

There is one potential problem.  Some malloc routines, if they were to
fail, would be recoverable.  In the current code, they may just unlock
any applicable locks and fail...but eventually they may be able to
recover fully.  By comparison, under ggz_malloc and friends there is no
way to avoid exiting after a failed malloc/realloc.

The chance of introducing bugs in this change is small but
non-negligible.  It will have to be tested closely.