[svn] Set room->spectating to 0 when we launch a game.
This value really needs to be tracked better: updated when we actually
join a table rather than when we try to.  Also ggzd should not rely on
the client to tell it whether it is spectating or not: doing so is why
this caused a bug in the first place.

(Bug: open two clients, and connect to a server.  With one client, launch
a game.  Then spectate it with the other.  Exit the game.  Then launch a
game with the second client.  The client and the table will now be
fubar.  This is fixed with this change, but a different client could
still mess things up.)