[svn] Implement a table_copy() function to make a copy of a table.  This copy can then be used as a snapshot without having to worry about locking.
Before this change, a snapshot was made just by saying *new_table =
*old_table (as opposed to new_table = table_copy(old_table)).  The
problem with this was that the seat and spectator data, which is stored
in dynamically allocated arrays, is not copied.  Therefore we can
(although I found this very hard to reproduce) end up accessing stale or
invalid data when we try to look at the "snapshot" later.

The only down side of this change is that making a snapshot is a bit
trickier - we have to make sure to call table_free on the snapshot when
we're done with it.