[svn] To help in debugging: ggz_error_msg_exit and ggz_error_sys_exit abort(), instead of just exiting.  ggz_error_msg and ggz_error_sys fork and abort, to leave core while the original process continues.
If core is disabled "ulimit -c 0", no core will be left and the behavior
will be virtually unchanged (at least on my platform).

Eventually something slightly more complicated may be better: we can
attach the core to gdb (if it exists) and mail a backtrace to ggz-dev.
THere's probably some tool out there for automating this.