[svn] Don't list 0 bots under BotsAllowed for any games.
There is a slight complication here: ggz_conf doesn't seem to allow empty
strings (it treats them as null and gives an error).  So rather than
leave an empty string for games that don't allow bots I've commented the
line out (which causes ggz_conf to not find it, and ggzd will use its
default value of "").