[svn] - in login screen, ask for game server hostname unless GGZ mode is enabled - prepare direct login functionality
This is intended to be compatible with AGS so the server can easily be
switched:
When connecting to the traditional Keepalive server, only one World is
available. When connecting to AGS, many more can be found there.

Of course, this still doesn't add any sense to the game itself, but that's not
the point here :-)