[svn] - ggzdmod gets a higher interface number too
For example, the old API of ggzdmod_set_module is not available any more.
Chances are high that the current API is relatively stable.
However, if we ever decide to keep an interface number, we must double-check
using 'nm' to see if non-static functions or variables have changed or not.