[svn] Add handling for a table chat (GGZ_CHAT_TABLE) at the server end.
Because of the shortcomings of the event system, this is not quite as
trivial as it might otherwise be.  We must first enqueue a table event,
then in the callback function for that event loop over all players and
spectators at the table and enqueue a player event to send them the chat.
This is commented on in the code (along with a possiblity for a better
way).