[svn] Changes to ggzcore to send GGZ data to the game client through ggzmod.
One protocol change was required: the ggzcore_game_init() function now
takes as an additional parameter the GGZServer* server variable.  I
also removed the obsolete function ggzcore_game_join.

However, the way ggzcore is structured is not ideal for having the GGZ
information be communicated to the game table.  Currently when the game
launch is done, the player's name is known and we know if they are
spectating.  But we will not generally know the player's seat number
(which is assigned by ggzd and sent back to us later) or what the
other seats at the table are (we may know if we join a table, but not
if we launch one).  In fact if the client calls the functions
ggzcore_room_join_table and ggzcore_game_launch out of order, we
might not even know this bit of information at the time the game client
is launched.

The solution to these problems may involve cleaning up ggzcore and
changing the API some more.