[svn] Improvement (hopefully :-) to acinclude.ggz.  Now each GGZ check takes two optional parameters.  For instance:
  AC_GGZ_LIBGGZ(action-if-found, action-if-not-found)

'action-if-found' will be evaluated if the specified module was found.
'action-if-not-found' will be evaluated if the module is not found.  This
is consistent with other autoconf macros.  If 'action-is-not-found' is
empty, then an error will be emitted if the module isn't found.  If
'action-if-not-found' is "ignore", then the module will simply be ignored
if it's not found.

This is not entirely backwards-compatible, since "ignore" used to be
the first parameter to the macro.  But this way is more consistent, in
my opinion.