[svn] Change GGZd so it no longer uses the "transit" waiting period when making a seat change at a table.  Now all seat changes are made directly by GGZd's table thread, who informs (rather than asks) the game server about it.
Since a single process is now in charge of authorizing and making changes,
it becomes possible for more complex transactions to be done without
changing the entire infrastructure.  See the ggz-dev posts on
"ggzd <-> ggzdmod <-> game server question".

Note that ggzdmod is left unchanged right now, although in the long run
this change will allow it to be substantially simplified.