[svn] Add support for table booting, and improve the code for desc and seat updates.
This code is currently very dangerous.  Locking is not done properly, so
if the table in question goes away while we're working on it we could
get some really nasty results.  There is a similar problem when dealing
with the player we're booting.  Fixing this will take some significant
code (or structural changes).

Booting a player is currently done just by starting them on a
GGZ_TRANSIT_LEAVE.  Unfortunately, the code doesn't know the difference
between a boot and a player <LEAVE>, and will send the player a <RESULT>
tag when the leave completes.  This is probably harmless for now.