[svn] Added handling for the RESEAT tag.
Doing this right entailed significant changes to the transit code,
including the unification of SPECTATOR and SEAT change transit events.
However this is offset by the lack of elegence of the new code (which may
be improved over time).

There is currently no 'reseat' UPDATE.  When a reseat happens, the update
sent consists of a 'leave' followed by a 'join'.  This will inevitably
cause problems for some clients.