[svn] When we're launching the game, we fork() and then exec().  After the fork, if there are any errors we should *not* call ggz_debug or any of its friends (including ggz_error_sys_exit()).  These may pass on their output to a handler function, which may not work properly since we've forked() off of the main program.
Instead, when this happens we send a ggzdmod log message.  This cannot be
done (easily) with ggzdmod_log, since the ggzdmod object we have is a
GGZ-side one and we're now on the game side.  Instead we access the IO
functions directly to send first a log message and then to change the
state to DONE.

This could end up being a useful debugging tool at some point, since if
the exec fails (for instance, because the command was not set correctly)
we will not get a useful ggzd.debug line about it.  But it would be more
useful if the game could send *error* messages to ggzd, independently of
the "debug" messages it currently sends.