[svn] Added client support for the <JOIN> and <LEAVE> tags sent by the server.
  <JOIN TABLE='5'/>
  <LEAVE REASON='boot' PLAYER='jdorje'/>

Now the client will no longer treat the <RESULT> tags (actions 'join' and
'leave') as indications that the player has joined/left the table.  It
will wait to see the appropriate <JOIN> or <LEAVE> tag.

Core clients will only be slightly affected: there is now some useful
event data for the GGZ_TABLE_JOIN and GGZ_TABLE_LEFT room events.

The information sent by ggzd is not currently checked - we just change the
GGZServer appropriately and then send it to the core client.  There's not
much we could do anyway...

Note that this change will break network compatabiliy.  Don't try to play
a game on an older server!