[svn] Fix a severe bug.
Until yesterday, when sending a <PLAYER> over the network (generally to
a different player), the player structure was not locked correctly.  There
is a hash_player_lookup in net.c, but then the lock is released before the
data is sent.  However player_get_type re-locked the structure before
determining the player type.

This sounds, good, but it is dangerous: once the lock is released, the
player could even log out and the player structure could be destroyed.  So
we could end up deadlocking or segfaulting when we try to access the
data later.

Yesterday, I changed this code so it makes a copy of the player structure
while it still has the lock.  This is the correct way to do things, AFAICT
(some FIXME comments in the code suggest it was the original intention,
and just wasn't completed).  But this lead to a deadlock when
get_player_type tried to access an invalid (copied) lock.

So the correct fix now is just to have get_player_type *not* acquire a
lock, and document this in the function's prototype.  Nobody else calls it,
so this won't break anything else.