[svn] Complete rewrite of the game server using the TTT server as a reference.
Hence supports as much of spectators and stats as TTT does, although the client has support for neither.

Synchronising hasn't been tested.

A rudimentry AI has been implemented. It either moves in the direction in which it should be escaping or else takes a random move.
It is very easy to beat :)