[svn] - add autonetwork() method to ggzmod python wrapper
This is mostly copied from the SDL games and certainly not the optimum.
But it frees the game developer from any unnecessary activity.
In python, the whole code is now reduced to:

	ret = ggzmod.autonetwork(networked, gamefd)
	if ret == 1 or ret == 3:
		network()
	if ret == 2 or ret == 3:
		ggzmod.dispatch()

(Using ret & 1, ret & 2 would be more elegant, but later on the return type
will turn into an object anyway.)