[svn] - remove obsolete acconfig.h, modify configure.ac accordingly - check for ggzcore if only telggz is selected - remove check for popt because it was only used by grubby once upon a time
There is a problem left:
If only GNOME 2.x is installed, the macro INSIDE_GNOME_COMMON is not defined,
which will always abort the configuration (no matter whether motd-editor is
included).
My guess is that motd-editor should be ported to GNOME 2.x so we can drop 1.x
support entirely, and have less problems.