[svn] - change QSocket to QSocketDevice - fix bug: don't try to read move information if move was invalid
The QSocket(Device) changes are really evil. They potentially break all applications
using QDataStream on them, like Kamikaze or Keepalive.
Previously, if not enough data was available, the socket blocked for a while,
whereas now it simply returns garbage.
I consider it to be a workaround until Qt gets sane again.
The affected versions are AFAIK Qt 3.2 (or 3.1?) through 3.3.