[svn] Allow libggz to compile and be used on systems that don't have syslog.h
A configure-time check is done for syslog.h.  If present, we just #include
it.  If it's not present, an enumeration containing the log types is
declared.  This must be done at configure-time so that the installed
ggz.h file can be entirely self-contained: we can't rely on the library
user following "good" autoconf practices.