[svn] Two of the three status bars (messagebar and statusbar) along the bottom of the main window were incorrectly showing resizer grips. They now no longer show the grip.
In addition, messagebar and statebar have been given a requested minimum width of 150 pixels. Before this commit, they were squashed into a few pixels by statusbar which is set to expand to fill the available space. This was making the information in statebar unreadable.

It should be noted that it does not appear as though statusbas (far left) and messagebar (middle) get used at all.