[svn] - Support for IPv6
This is riq's old patch (4 years! oh my!) with some slight changes to make
it work, including static methods so we don't have to bump the interface.
Since ggzcommunity.org is accessible via IPv6, and it works, this should
be made available to all users.

Of course, IP ban protection in ggzd (from clients using IPv6) is a
different matter.