[svn] Preliminary abandoned-seats patch.
In principle when a player leaves a running game, his seat becomes
abandoned.  He may later rejoin and reclaim his seat, or after a while
he may be given a forfeit.  For the most part this should be controlled
by the game or by ggzd according to the game's options.

This patch introduces the preliminary changes for this.  A new seat type
GGZ_SEAT_ABANDONED is added, and basic changes are made to support it.
What is missing is ggzd and ggzdmod support for toggling between player,
abandoned, and open seats.  There's also no forfeit mechanism yet (although
the game may do this by hand if desired).

This patch came from the playground; see patches/abandoned.diff for its
history.