[svn] Recent changes to easysock do not appear to be supported under winsock on Win32. This means winsock2 must be used. This update changes libggz to use winsock2.
easysock.c includes an extra header file, ws2tcpip.h, which I have not checked for in configure.ac on the reasoning that it is part of the winsock2 package and so HAVE_WINSOCK2_H should imply ws2tcpip.h exists as well. Shout at me for this if you want.

easysock.c makes use of the gai_strerror() function which is not implemented in current mingw builds. See bug 1188891 at mingw or this url: https://sourceforge.net/tracker/?func=detail&atid=102435&aid=1188891&group_id=2435 It has been fixed in CVS so should appear in a release soon hopefully.