[svn] - do *not* link against ggzmod
I read the libtool docs again, and it clearly says:
A library A which is only used by a library B the programme depends on, but
not by the programme itself, should only be added to the LIBADD of said
library B, but not to the LDADD of the programme.

And, it seems to work. If it ever doesn't work, something else is wrong.