[svn] - rename all ggzmod_XXX functions to ggzmod_ggz_XXX
I never thought this was necessary. BUT:
from ggzcore, both ggzmod and ggzmod-ggz have the same "distance", so it
is confused at which function symbol to resolve... and naturally goes for
the wrong one.
Ideally, it would know that a statically linked lib should get a higher
score than a dynamically linked one. Maybe this can even be configured.

This was not so much a problem for normal games and core clients, but very
much for embedded games which use both ggzmod and ggzcore.