[svn] - improved launch dialog - prepare possibility to add grubbies as "named bots" too
Grubbies can only be added if they're in the current room, as they
are not registered with the game server (anybody can run one).
The problem for KGGZ is to find out which games are supported by
grubby.
Then, if grubby sees a table launch where there's a seat reserved
for him, he should join automatically, the the "/join my game" command
becomes superfluous.