[svn] - move log handler to ggz-side of ggzdmod interface - correctly implement named bots by passing ggzdmod the .dsc file name
Both changes break compatibility again and require another recompilation!

Actually it'd be more "clean" to pass the named bots to ggzdmod. But since
other additions might follow, and we don't constantly want to modify the
ggzd<->ggzdmod protocol with all the wrapping involved, it's easier to
let ggzdmod-game do some of the handling itself.