[svn] Add previously missing function ggzmod_is_ggz_mode.  This returns TRUE if the program is being run in GGZ mode (as determined by environment variables). With this function a program can determine if it should exit with an error or if it should run in non-ggz mode.
One problem is that with embedded GGZ, once the embedded GGZ begins GGZ mode
is entered and is not left.  This is probably a bug elsewhere in ggzmod.  It's
not clear in this case whether the ggz mode environment variable should be
left unset, or if it should be cleared when embedded GGZ is complete.