[svn] - fix the ${datadir} fix
Finally the difference between this and ${datarootdir} became clear to me.
On some systems, game data is installed into /usr/share/games.
This does not necessarily apply to GGZ because we're more than a lousy game
but ${datadir} is still correct.

The fix introduces AC_GGZ_UNQUOTEDPATH, which hopefully turns out to be working.