[svn] Remove unhelpful/spam message "A room was removed!"
This message could be useful for debugging, but it needs to give at least the
room name.  It is not a good way to give this information to users since it
cannot be translated or presented outside of the chatline - for that we need
the core client to recognize the removal as it happens and just tell the
user directly.