[svn] Added the PERMS_TABLE_PRIVMSG permission, which allows a player to send private messages whilst sat at a table.
This means:

User with perm at table, user without not at table -> privmsgs allowed both ways
User with perm at table/not at table, user without at table -> privmsgs not allowed

It may also be sensible to allow messages from a user with the permission to a user without the permission who is at a table, but not the other way. I have not included this for the moment as it is potentially more confusing for the unprivileged player.