[svn] Start of completing the mysql database functions. I've moved the sql escaping functions to the common database file from the postgres one so that the mysql functions can take advantage of them.
The functions that I've implemented are essentially copies of the functions in the postgres copy, but with mysql-isms instead. I've copied the pthread_mutex_*() function calls as they are used in the rest of the stats functions, but it would be a good idea is someone else could cast their eye over it all to make sure that it is correct.