[svn] Fixed a load of malloc/free -> ggz_malloc/ggz_free calls related to libggz/src/security functions. Fixed a few minor memory leaks.
I'm still getting about 68k of memory leaked just by starting a server and then quiting it (other than the small chunks leaked in 3rd party libs that is). Likewise if I kill a server with SIGHUP there is a smaller amount of memory leaked. It looks to me as though it's caused in event_*_enqueue() but I haven't got that far into it yet.