[svn] - first version of a synchronisation trigger which runs in the database backend
Since it was written twice (accidentally deleted the first version), I didn't
really progress to the point where synchronisation works. This is merely a
first shot at testing the compilation, installation and activation of a trigger
which is purely informative.

For the time being, peers.conf needs to be copied to /tmp (yuck!)