[svn] - replace all -isystem with simple -I
It seems that only a subset of modern GCCs (maybe early 3.x only?) had to be silenced with -isystem. The effect is not quite the same,
as -isystem always gets used after all -I flag.
Even if we still had to use -isystem for GCC, we would use -I as a fallback for the other compilers.