[svn] - dynamic room changes: handly room updates
Until now, we handled either additions (not present -> parse) or removals (present -> remove) for inotify-based updates. We did not handle updates (present -> remove & parse), which resulted in updates causing a series of removals and additions.

When players are in the room during the updates, a new room with the same name will co-exist for some time, while the old one will be locked and people can only leave, not join. Given a proper representation in the core clients, this should be an intuitive process.