[svn] - addition of (not yet used) function to cut Unicode strings
While adding it I also found out that http://bugs.icu-project.org/trac/ticket/5638
was closed saying that returning E_BUFFER_OVERFLOW_ERROR is the expected behaviour
for the «pre-flight» checks for the number of bytes needed for malloc.
Obviously, such an opinion and the use of an Excel sheet by IBM to track the bugs
of ICU do not really come as a surprise anymore. The rest of our code needs to be
adapted to this «fix», though.