[svn] Fix a major bug with leaving games.  Prior to this fix the game_tag was never reset, as the socket value has already been reset to -1 at this point.
There is still a bug present though in ggzcore.  Each time a game is
left it triggers the GGZ_TABLE_LEFT event twice, once when ggzd sends
the LEAVE packet and a second time when the game client exits and the
connection is cut; the second one also incorrectly generates a
GGZ_TABLE_LAUNCH_FAIL event.  I leave the assertions in to help us sort
out this mess but note they make ggz-gtk unusable right now.