[svn] - display player faves on the public information page
There are two issues here: First of all, we want to allow (with conservative defaults)
which information goes public at all. Second, from a technical perspective, we don't
handle rooms at all in the web interface except for the dynamic rooms menu for admins/hosts.
But it would certainly be useful to distinguish "TTT Beginners" from "TTTL33TXXL" and so forth.